requirements.txt
obs-websocket-py>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
python-dotenv>=1.0.0